Don't open temp file if __input is absent
This fixes a logic bug: due to the truthiness of the empty string,
the temp file is always created, and the shell always called with
a file redirect.

While this is unlikely to change the result of a program, it
could, and the implication of checking for args.input is that this
was unintended behavior.
 On branch fix-input-logic
	modified:   sh.lua